Improve 'open' command' shortcuts refresh concurrency
- By default, 'open' command shortcuts are loaded on startup in the main
thread on win32 and asynchronously on Linux/OSX. Tests shows that on
Windows it's actually faster to initially refresh the shortcuts in main
thread rather than asynchronously in separate thread. Background updates
are still performed in a thread.
This is configurable in config.
- Avoid re-entrance of the background reloader thread for background
update of 'open' command shortcuts
- Introduced debouncing logic (4 seconds) to 'open' command background
shortcuts reload to avoid excessive reload triggering on massive changes
to monitored directories.